src/main/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
redis[asyncio]==5.0.1
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
python-multipart==0.0.6
